+++ src/components/cart/CartSummary.css
@@ -1,0 +1,85 @@
+.cart-page__summary {
+  flex-shrink: 0;
+  width: 22rem;
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+  padding: 2rem;
+  border-radius: var(--radius-lg);
+  background-color: var(--primary-bg);
+  border: 1px solid var(--gray-100);
+  box-shadow: 0 20px 40px rgba(15, 23, 42, 0.12);
+}
+
+.cart-page__summary-title {
+  margin: 0;
+  font-size: var(--text-xl);
+  font-family: var(--font-heading);
+  color: var(--gray-900);
+}
+
+.cart-page__summary-meta {
+  margin: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.cart-page__summary-row {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 1rem;
+}
+
+.cart-page__summary-label {
+  font-size: var(--text-sm);
+  color: var(--gray-500);
+}
+
+.cart-page__summary-value {
+  font-size: var(--text-base);
+  font-weight: var(--font-medium);
+  color: var(--gray-700);
+}
+
+.cart-page__summary-value--price {
+  font-weight: var(--font-bold);
+  color: var(--gray-900);
+}
+
+.cart-page__summary-total {
+  display: flex;
+  align-items: baseline;
+  justify-content: space-between;
+  border-radius: var(--radius-md);
+}
+
+.cart-page__summary-total-label {
+  font-size: var(--text-base);
+  color: var(--gray-700);
+}
+
+.cart-page__summary-total-value {
+  font-size: var(--text-title-xl);
+  font-weight: 700;
+  color: var(--primary-500);
+}
+
+.cart-page__summary-button {
+  width: 100%;
+  padding: 0.95rem 1.25rem;
+  border: none;
+  border-radius: var(--radius-md);
+  background-color: var(--primary-500);
+  color: var(--gray-00);
+  font-size: var(--text-base);
+  font-weight: var(--font-bold);
+  cursor: pointer;
+  transition: background-color 0.2s ease, box-shadow 0.2s ease;
+}
+
+.cart-page__summary-button:hover {
+  background-color: var(--primary-400);
+  box-shadow: 0 16px 26px rgba(26, 145, 209, 0.25);
+}
